+++ src/common/resolver-args.ts
@@ -1,0 +1,4 @@
+export interface PaginationArgs {
+  limit?: number;
+  offset?: number;
+}
