+++ src/modules/app.module.ts
@@ -1,16 +1,15 @@
-import { mergeTypeDefs } from "@graphql-tools/merge";
+import { mergeTypeDefs, mergeResolvers } from "@graphql-tools/merge";
+
 import { bandsTypeDefs } from "./bands/bands.schema.js";
 import { userTypeDefs } from "./users/user.schema.js";
+import { genresTypeDefs } from "./genres/genres.schema.js";
 
 import { userResolvers } from "./users/user.resolvers.js";
 import { bandsResolvers } from "./bands/bands.resolvers.js";
-import { mergeResolvers } from "@graphql-tools/merge";
+import { genresResolvers } from "./genres/genres.resolvers.js";
 
 import { UsersAPI, UserService } from "./users/user.service.js";
 import { BandsAPI, BandsService } from "./bands/bands.service.js";
-
-import { genresTypeDefs } from "./genres/genres.schema.js";
-import { genresResolvers } from "./genres/genres.resolvers.js";
 import { GenresAPI, GenresService } from "./genres/genres.service.js";
 
 export const typeDefs = mergeTypeDefs([userTypeDefs, genresTypeDefs, bandsTypeDefs]);
